221129_1032
캣츠아이 레벨 적용
로스러크 레벨 적용
매직볼 레벨 적용

콘솔 매니저 추가

-레벨 시스템 추가
-경험치 시스템 추가
-드롭 노란별 추가
-드롭 푸른별 추가
-드롭 초록별 추가
-드롭 붉은별 추가
-드롭 하얀별 추가

-장비 "리본" 추가
-장비 "마법의 모자" 추가
-장비 "돋보기" 추가
-장비 "마법석" 추가
-장비 "얼어붙은 이름표" 추가
-장비 "가속 포탈" 추가
-장비 "프로텍트볼" 추가
-장비 "밤송이" 추가
-장비 "액상 사료" 추가
-장비 "푸른 별 조각" 추가

-치트 패널 생성
-치트 프리뷰 생성
-치트 입력 생성
-버튼형 치트 생성
-버튼형 치트 별명 생성

-포커싱 아웃 일시정지 생성

+++ Assets/Script/Bullet/Bullet.cs
@@ -6,7 +6,7 @@ public class Bullet : MonoBehaviour
 {
     public float speed;
     public float dmg;
-    public int knockback;
+    public float knockback;
     protected virtual void Awake()
     {
         GetComponent<BoxCollider2D>().isTrigger = true;
